--- components/GNB/Alarm.tsx
@@ -1,7 +1,0 @@
-export default function Alarm() {
-  return (
-    <button>
-      <img src="/icon/icon-alarm.svg" className="mo:hidden" alt="알림 아이콘" />
-    </button>
-  );
-}
