--- pages/test/GNB/index.tsx
@@ -1,9 +1,0 @@
-import GNB from '@/components/GNB/GNB';
-
-export default function GNBtest() {
-  return (
-    <div>
-      <GNB />
-    </div>
-  );
-}
