+++ pages/test/Headers/index.tsx
@@ -1,0 +1,9 @@
+import Headers from '@/components/Headers/Headers';
+
+export default function HeaderTest() {
+  return (
+    <div>
+      <Headers />
+    </div>
+  );
+}
